aoai/ptu-resource-automation/requirements.txt
azure-identity
azure-mgmt-automation
azure-mgmt-authorization
azure-mgmt-resource
python-dotenv